added more songs, buttons neighbours automated
so initially we had to manually add the focus neighbours for the
buttons. That obviously makes it a pain in the ass to deal with if we
have a lot of menu items (for example I plan on adding a songlist where
we can manually pick them). This commit deals with that and
automatically sets the buttons up for us