vol1 locks if it can't find the v1s1 scene
plan is to make the client a separate tool, and music itself another. So
when I drop the volume 2, people will just have to add vol2 to the .pkc
file. But I have to fix the client buttons so it works that way without
requiring the users to update the client