added icons for some inventory items
biggest problem here was the CRT shader being too harsh in this case.
When you open the inventory with CRT shader on, everything lighter will
blow completely out and hide the extra details. It took me a while to
figure it out, but I managed to write some extra shader to apply
specifically for the icons. What the shader does is just make the icons
darker, and add some shadows around the items